safari-600.7-branch/LayoutTests/fast/dom/HTMLTableElement/empy-table-cell-with-background-color.html

<!DOCTYPE html>
<html>
    <head>
        <script src="../../../resources/js-test-pre.js"></script>
        <script>
            function runTest() {
                description('Tests that the table cell width is set properly even when the table cell has no child and its table has width defined');

                shouldBe("document.getElementById('nochildTd').offsetWidth","750");
                isSuccessfullyParsed();
            }
        </script>
    </head>
    <body onload="runTest()">
        <table  width="750px" border="0" cellpadding="0" cellspacing="0">
            <tr>
                <td id="nochildTd" height="1px" bgcolor=#000000></td>
            </tr>
        </table>
    </body>
</html>
