LayoutTests/fast/dom/HTMLTableElement/empy-table-cell-with-background-color-expected.txt

Tests that the table cell width is set properly even when the table cell has no child and its table has width defined

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.getElementById('nochildTd').offsetWidth is 750
PASS successfullyParsed is true

TEST COMPLETE
